delete functionality works

+++ src/lib/components/DeleteButton.tsx
@@ -11,6 +11,8 @@ interface iDeleteButtonProps extends React.PropsWithChildren {
 export const DeletButton: React.FC<iDeleteButtonProps> = (props) => {
   const [isDeleting, setIsDeleting] = useState<boolean>(false);
   const [deleteStr, setDelStr] = useState<string>("");
+  const [showDelErr, setShowDelErr] = useState<boolean>(false);
+  const [loading, setLoading] = useState<boolean>(false);
   const dispatch = useAppDispatch();
 
     function handleDeleteClick(){
@@ -20,20 +22,34 @@ export const DeletButton: React.FC<iDeleteButtonProps> = (props) => {
 
     const handleDeleteSubmission = async (e:FormEvent) => {
         e.preventDefault();
-        try{
-            const idToDel = await axios.put(`api/questions/delete/${props._id}`, {id:props._id});
-            // console.log(idToDel);
-            dispatch(removeQuestion(idToDel.data.id))
+        setLoading(true);
+        if(deleteStr === "DELETE"){
+          try{
+              const idToDel = await axios.put(`api/questions/delete/${props._id}`, {id:props._id});
+              // console.log(idToDel);
+              dispatch(removeQuestion(idToDel.data.id))
+          }
+          catch(err:any){
+              console.log(err);
+          }
+          finally{
+            setLoading(false);
+          }
         }
-        catch(err:any){
-            console.log(err);
+        else{
+          setShowDelErr(true);
+          setTimeout(() => {
+            setShowDelErr(false);
+            setLoading(false);
+          }, 3000);
         }
     }
 
   return (
     <div>
         <button className="bg-red-700 text-slate-300 font-semibold rounded-full p-3 relative hover:bg-red-800 hover:text-slate-100 transition-all duration-200 disabled:bg-red-400 disabled:text-black" onClick={handleDeleteClick} disabled={isDeleting}>DELETE</button>
-        <div className={`${isDeleting ? "visible":"invisible"} relative bottom-0 bg-slate-100 py-4`}>
+        {isDeleting ? 
+          <div className="bg-slate-100 py-4">
             If sure type DELETE into the dialogue box and press the delete button
             <form onSubmit={handleDeleteSubmission}>
                 <input name='delete' value={deleteStr} onChange={(e)=>setDelStr(e.target.value)} />
@@ -43,9 +59,17 @@ export const DeletButton: React.FC<iDeleteButtonProps> = (props) => {
                     setIsDeleting(false);
                     }} className='bg-indigo-300 text-stone-600 font-semibold rounded-full p-3 hover:bg-indigo-800 hover:text-stone-800'>Cancel</button>
                 <br />
-                <button type='submit' className='bg-red-700 text-slate-300 font-semibold rounded-full p-3 hover:bg-red-800 hover:text-slate-100'>Submit</button>
+                <button disabled={loading} type='submit' className='disabled:bg-opacity-30 bg-red-700 text-slate-300 font-semibold rounded-full p-3 hover:bg-red-800 hover:text-slate-100'>Submit</button>
             </form>
+          </div>
+        :
+          ""
+        }
+        {showDelErr ? 
+        <div className='text-xs text-red-800 bg-slate-100'>
+          Please type {"'DELETE'"} or click the {"'Cancel'"} button
         </div>
+        : ""}
     </div>
   )
 }
